pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "liteagent"
version = "0.1.0"
description = "轻量级的 AI Agent 框架"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Jiayu Xu", email = "your.email@example.com"},
]
maintainers = [
    {name = "Jiayu Xu", email = "your.email@example.com"},
]
keywords = ["ai", "agent", "llm", "framework", "assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.8"
dependencies = [
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.350",
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
]
all = [
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/JiayuXu0/LiteAgent"
Repository = "https://github.com/JiayuXu0/LiteAgent.git"
Issues = "https://github.com/JiayuXu0/LiteAgent/issues"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/liteagent"]

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/node_modules", "**/__pycache__", "build", "dist"]
pythonVersion = "3.8"
pythonPlatform = "All"
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false
reportOptionalMemberAccess = false
reportGeneralTypeIssues = true
reportOptionalCall = false

[tool.ruff]
target-version = "py38"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle 错误
    "F",      # pyflakes
    "W",      # pycodestyle 警告
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "RUF",    # Ruff 特定规则
]
ignore = [
    "E501",   # 行太长（由 line-length 控制）
    "B008",   # 函数调用中的可变参数默认值
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["liteagent"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=src/liteagent",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: 标记为慢速测试",
    "integration: 集成测试",
    "unit: 单元测试",
]

[tool.coverage.run]
source = ["src/liteagent"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pyright>=1.1.403",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.12.8",
]
